requirements.txt
pandas
datasets
nltk
sentence-transformers
langchain
langchain-core
langchain-community
langchain-google-genai
faiss-cpu
ragas
streamlit
python-dotenv